Add report processing pipeline
This is the meat of the collector: an HTTP handler that listens for POST
uploads, and parses the JSON array of reports in the payload.  The
generic ReportProcessor interface will allow us to customize and
configure how exactly the reports are handled.  Right now, there aren't
any interesting processors, but eventually we'll add some filters and a
variety of output processors for publishing the reports to external
logging or monitoring systems.

This patch also adds a `nel-collector` command.  It's currently not
configurable; it listens on localhost:8080 and prints out a summary of
each report to stdout.  Eventually it will be much more interesting.