Add client IP and user-agent to report batches
Just like collection time, this information only needs to be recorded
for batches, not for the individual reports within a batch.

This also gives us all the information we need to make the report dumper
use Apache's CLF format.  Er, well, _almost_ all the information; we
should probably update the spec to include the HTTP verb...