Add testdata directory
This patch moves the JSON report content that we use in our test cases
into a testdata directory.  The expected results are now stored in
golden files in the same directory.  The test suite now accepts an
`--update` flag that will update the golden files with the current
actual output.