Allow multiple named annotations
Instead of having a single `Annotation` field on reports and batches,
there's now a separate map-like class that we embed in each of them,
which lets you define several independent named annotations.  This will
make is easier to different "groups" of processors that operate on
different annotations, without having them conflict with each other.