Remove more boilerplate in pipeline tests
This is a much nicer helper class for testing report processor
pipelines.  It automatically finds all of our test input payloads, runs
the pipeline against all of them, and compares the actual output (which
the pipeline should store into a the `TestResult` annotation) with the
contents of a golden file.  The testdata directory structure is now more
nested, making it (hopefully) more obvious which files belong to which
test cases.

Bug: #19