Use configuration file to define report processors
We're going to use TOML as the syntax.  You'll add a top-level
`processor` clause for each processor you want to instantiate.  The
`type` element determines which kind of processor to instantiate, and
the processor's loader function can use any other elements to configure
the new processor.

This patch also adds a new test helper function that instantiates the
pipeline to be tested using a configuration string.  Each of the
per-processor test cases is now really clean and simple.